services/memory-engine/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
openai==1.2.4
chromadb==0.4.18
python-dotenv==1.0.0
loguru==0.7.2
httpx==0.25.1
numpy==1.26.1
tenacity==8.2.3
pytest==7.4.3
